broken extract gameloop and ecsmanagment
